Cast all std::vector.size() returns in the RVC service area delegate to uint32_t. The service area server limits the size of these lists to 255.

+++ examples/rvc-app/rvc-common/src/rvc-service-area-delegate.cpp
@@ -109,14 +109,14 @@ bool RvcServiceAreaDelegate::IsSupportedLocationsChangeAllowed()
 
 uint32_t RvcServiceAreaDelegate::GetNumberOfSupportedLocations()
 {
-    return mSupportedLocations.size();
+    return (uint32_t)mSupportedLocations.size();
 }
 
 bool RvcServiceAreaDelegate::GetSupportedLocationByIndex(uint32_t listIndex, LocationStructureWrapper & aSupportedLocation)
 {
     bool ret_value = false;
 
-    if (listIndex < mSupportedLocations.size())
+    if (listIndex < (uint32_t)mSupportedLocations.size())
     {
         aSupportedLocation = mSupportedLocations[listIndex];
         ret_value          = true;
@@ -134,7 +134,7 @@ bool RvcServiceAreaDelegate::GetSupportedLocationById(uint32_t aLocationId, uint
     bool ret_value = false;
 
     listIndex         = 0;
-    uint32_t listSize = mSupportedLocations.size();
+    uint32_t listSize = (uint32_t)mSupportedLocations.size();
 
     while (listIndex < listSize)
     {
@@ -159,11 +159,11 @@ bool RvcServiceAreaDelegate::AddSupportedLocation(const LocationStructureWrapper
     // etc.
 
     // Double-check list size to ensure there no memory issues.
-    if (mSupportedLocations.size() < kMaxNumSupportedLocations)
+    if ((uint32_t)mSupportedLocations.size() < kMaxNumSupportedLocations)
     {
         // not sorting list, number of locations normally expected to be small, max 255
         mSupportedLocations.push_back(newLocation);
-        listIndex = mSupportedMaps.size() - 1; // new element is last in list
+        listIndex = (uint32_t)mSupportedMaps.size() - 1; // new element is last in list
         ret_value = true;
     }
     else
@@ -200,7 +200,7 @@ bool RvcServiceAreaDelegate::ClearSupportedLocations()
 {
     bool ret_value = false;
 
-    if (mSupportedLocations.size() > 0)
+    if ((uint32_t)mSupportedLocations.size() > 0)
     {
         mSupportedLocations.clear();
         ret_value = true;
@@ -219,14 +219,14 @@ bool RvcServiceAreaDelegate::IsSupportedMapChangeAllowed()
 
 uint32_t RvcServiceAreaDelegate::GetNumberOfSupportedMaps()
 {
-    return mSupportedMaps.size();
+    return (uint32_t)mSupportedMaps.size();
 }
 
 bool RvcServiceAreaDelegate::GetSupportedMapByIndex(uint32_t listIndex, MapStructureWrapper & aSupportedMap)
 {
     bool ret_value = false;
 
-    if (listIndex < mSupportedMaps.size())
+    if (listIndex < (uint32_t)mSupportedMaps.size())
     {
         aSupportedMap = mSupportedMaps[listIndex];
         ret_value     = true;
@@ -243,7 +243,7 @@ bool RvcServiceAreaDelegate::GetSupportedMapById(uint8_t aMapId, uint32_t & list
     bool ret_value = false;
 
     listIndex         = 0;
-    uint32_t listSize = mSupportedMaps.size();
+    uint32_t listSize = (uint32_t)mSupportedMaps.size();
 
     while (listIndex < listSize)
     {
@@ -268,11 +268,11 @@ bool RvcServiceAreaDelegate::AddSupportedMap(const MapStructureWrapper & newMap,
     // etc.
 
     // Double-check list size to ensure there no memory issues.
-    if (mSupportedMaps.size() < kMaxNumSupportedMaps)
+    if ((uint32_t)mSupportedMaps.size() < kMaxNumSupportedMaps)
     {
         // not sorting list, number of locations normally expected to be small, max 255
         mSupportedMaps.push_back(newMap);
-        listIndex = mSupportedMaps.size() - 1; // new element is last in list
+        listIndex = (uint32_t)mSupportedMaps.size() - 1; // new element is last in list
     }
     else
     {
@@ -307,7 +307,7 @@ bool RvcServiceAreaDelegate::ClearSupportedMaps()
 {
     bool ret_value = false;
 
-    if (mSupportedMaps.size() > 0)
+    if ((uint32_t)mSupportedMaps.size() > 0)
     {
         mSupportedMaps.clear();
         ret_value = true;
@@ -321,14 +321,14 @@ bool RvcServiceAreaDelegate::ClearSupportedMaps()
 
 uint32_t RvcServiceAreaDelegate::GetNumberOfSelectedLocations()
 {
-    return mSelectedLocations.size();
+    return (uint32_t)mSelectedLocations.size();
 }
 
 bool RvcServiceAreaDelegate::GetSelectedLocationByIndex(uint32_t listIndex, uint32_t & aSelectedLocation)
 {
     bool ret_value = false;
 
-    if (listIndex < mSelectedLocations.size())
+    if (listIndex < (uint32_t)mSelectedLocations.size())
     {
         aSelectedLocation = mSelectedLocations[listIndex];
         ret_value         = true;
@@ -345,11 +345,11 @@ bool RvcServiceAreaDelegate::AddSelectedLocation(uint32_t aLocationId, uint32_t 
     // etc.
 
     // Double-check list size to ensure there no memory issues.
-    if (mSelectedLocations.size() < kMaxNumSelectedLocations)
+    if ((uint32_t)mSelectedLocations.size() < kMaxNumSelectedLocations)
     {
         // not sorting list, number of locations normally expected to be small, max 255
         mSelectedLocations.push_back(aLocationId);
-        listIndex = mSelectedLocations.size() - 1; // new element is last in list
+        listIndex = (uint32_t)mSelectedLocations.size() - 1; // new element is last in list
     }
     else
     {
@@ -364,7 +364,7 @@ bool RvcServiceAreaDelegate::ClearSelectedLocations()
 {
     bool ret_value = false;
 
-    if (mSelectedLocations.size() > 0)
+    if ((uint32_t)mSelectedLocations.size() > 0)
     {
         mSelectedLocations.clear();
         ret_value = true;
@@ -378,14 +378,14 @@ bool RvcServiceAreaDelegate::ClearSelectedLocations()
 
 uint32_t RvcServiceAreaDelegate::GetNumberOfProgressElements()
 {
-    return mProgressList.size();
+    return (uint32_t)mProgressList.size();
 }
 
 bool RvcServiceAreaDelegate::GetProgressElementByIndex(uint32_t listIndex, Structs::ProgressStruct::Type & aProgressElement)
 {
     bool ret_value = false;
 
-    if (listIndex < mProgressList.size())
+    if (listIndex < (uint32_t)mProgressList.size())
     {
         aProgressElement = mProgressList[listIndex];
         ret_value        = true;
@@ -403,7 +403,7 @@ bool RvcServiceAreaDelegate::GetProgressElementById(uint32_t aLocationId, uint32
     bool ret_value = false;
 
     listIndex         = 0;
-    uint32_t listSize = mProgressList.size();
+    uint32_t listSize = (uint32_t)mProgressList.size();
 
     while (listIndex < listSize)
     {
@@ -428,11 +428,11 @@ bool RvcServiceAreaDelegate::AddProgressElement(const Structs::ProgressStruct::T
     // etc.
 
     // Double-check list size to ensure there no memory issues.
-    if (mProgressList.size() < kMaxNumProgressElements)
+    if ((uint32_t)mProgressList.size() < kMaxNumProgressElements)
     {
         // not sorting list, number of locations normally expected to be small, max 255
         mProgressList.push_back(newProgressElement);
-        listIndex = mProgressList.size() - 1; // new element is last in list
+        listIndex = (uint32_t)mProgressList.size() - 1; // new element is last in list
     }
     else
     {
@@ -457,7 +457,7 @@ bool RvcServiceAreaDelegate::ClearProgress()
 {
     bool ret_value = false;
 
-    if (mProgressList.size() > 0)
+    if ((uint32_t)mProgressList.size() > 0)
     {
         mProgressList.clear();
         ret_value = true;
